[CLI] Move and rename files for `frontity create`
