+++ packages/frontity/package.json
@@ -11,7 +11,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "bin": {
-    "frontity": "dist/cli.js"
+    "frontity": "dist/cli/index.js"
   },
   "repository": {
     "type": "git",
@@ -24,8 +24,8 @@
     "access": "public"
   },
   "scripts": {
-    "start": "node dist/cli.js",
-    "dev": "ts-node src/cli.ts",
+    "start": "node dist/cli/index.js",
+    "dev": "ts-node src/cli/index.ts",
     "ts": "ts-node",
     "test": "jest --watch",
     "test:ci": "jest --colors",
